requirements.txt
# ── UI & app ───────────────────────────────────────────────────────────────
streamlit==1.44.1

# ── Media & video ──────────────────────────────────────────────────────────
moviepy==1.0.3
imageio==2.34.1
imageio-ffmpeg==0.5.1
Pillow==10.4.0
numpy==1.26.4
scipy==1.13.0

# ── TTS ────────────────────────────────────────────────────────────────────
edge-tts==6.1.12
pyttsx3==2.98

# ── LLM providers ──────────────────────────────────────────────────────────
openai==1.30.1
anthropic==0.25.0
httpx==0.27.0
requests==2.32.3

# ── System utilities ───────────────────────────────────────────────────────
psutil==5.9.8
jsonschema==4.22.0
